requirements.txt
psycopg2-binary
razorpay
python-dotenv
qrcode
pillow
python-telegram-bot
asyncpg>=0.29.0
fastapi
uvicorn
asyncpg
quart
hypercorn
